@@ -10,7 +10,7 @@
     "start": "npm run build && node dist/index.js",
     "start:dev": "nodemon --inspect=0.0.0.0:9229 --exec babel-node src/index.js",
     "heroku-postbuild": "npm run reverse && npm run migration && npm run seed",
-    "lint": "eslint src/ --fix",
+    "lint": "eslint src/ --fix . && echo 'Lint complete.'",
     "migration": "sequelize db:migrate",
     "seed": "sequelize db:seed:all",
     "reverse": "sequelize db:migrate:undo:all",
@@ -18,7 +18,7 @@
     "coverage": "nyc npm test && npm run generate-lcov && npm run coveralls-coverage && npm run codeclimate-coverage",
     "codeclimate-coverage": "codeclimate-test-reporter < lcov.info",
     "generate-lcov": "nyc report --reporter=text-lcov > lcov.info",
-    "test": "cross-env NODE_ENV=test mocha --require @babel/register src/test/**.test.js --timeout 20000 --exit"
+    "test": "cross-env NODE_ENV=test npm run heroku-postbuild && nyc mocha --require @babel/register src/test/**.test.js --timeout 20000 --exit"
   },
   "author": "Andela Simulations Programme",
   "license": "MIT",
@@ -40,6 +40,8 @@
     "mongoose-unique-validator": "^2.0.1",
     "morgan": "^1.9.1",
     "passport": "^0.4.0",
+    "passport-facebook": "^3.0.0",
+    "passport-google-oauth2": "^0.2.0",
     "passport-local": "^1.0.0",
     "pg": "^7.12.1",
     "pg-hstore": "^2.3.3",
@@ -47,7 +49,8 @@
     "sequelize": "^5.15.0",
     "sequelize-cli": "^5.5.0",
     "swagger-ui-express": "^4.0.7",
-    "underscore": "^1.9.1"
+    "underscore": "^1.9.1",
+    "uuid": "^3.3.3"
   },
   "devDependencies": {
     "@babel/cli": "^7.5.5",
@@ -58,12 +61,18 @@
     "chai-http": "^4.3.0",
     "concurrently": "^4.1.2",
     "cross-env": "^5.2.0",
-    "eslint": "^6.1.0",
+    "eslint": "^6.2.1",
     "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-config-prettier": "^6.1.0",
     "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-prettier": "^3.1.0",
+    "eslint-plugin-react": "^7.14.3",
+    "eslint-plugin-react-hooks": "^2.0.1",
     "expect.js": "^0.3.1",
     "mocha": "^6.2.0",
-    "nodemon": "^1.19.1"
+    "nodemon": "^1.19.1",
+    "nyc": "^14.1.1",
+    "prettier": "^1.18.2"
   },
   "engines": {
     "node": "10.0.0"
