@@ -1,83 +1,25 @@
-const mongoose = require('mongoose');
-const router = require('express').Router();
-const passport = require('passport');
+import { Router } from 'express';
+import passport from 'passport';
+import passportSetup from '../../config/passport'; // eslint-disable-line no-unused-vars
+import SocialMediaLogin from '../../controllers/socialMediaLogin';
 
-const User = mongoose.model('User');
+const { googleLogin, facebookLogin } = SocialMediaLogin;
 
-router.get('/user', (req, res, next) => {
-  User.findById(req.payload.id)
-    .then((user) => {
-      if (!user) {
-        return res.sendStatus(401);
-      }
-      return res.json({ user: user.toAuthJSON() });
-    })
-    .catch(next);
-});
+const router = Router();
 
-router.put('/user', (req, res, next) => {
-  User.findById(req.payload.id)
-    .then((user) => {
-      if (!user) {
-        return res.sendStatus(401);
-      }
+router.get(
+  '/users/google/login',
+  passport.authenticate('google', {
+    scope: ['profile', 'email']
+  })
+);
+router.get(
+  '/users/facebook/login',
+  passport.authenticate('facebook')
+);
 
-      // only update fields that were actually passed...
-      if (typeof req.body.user.username !== 'undefined') {
-        user.username = req.body.user.username;
-      }
-      if (typeof req.body.user.email !== 'undefined') {
-        user.email = req.body.user.email;
-      }
-      if (typeof req.body.user.bio !== 'undefined') {
-        user.bio = req.body.user.bio;
-      }
-      if (typeof req.body.user.image !== 'undefined') {
-        user.image = req.body.user.image;
-      }
-      if (typeof req.body.user.password !== 'undefined') {
-        user.setPassword(req.body.user.password);
-      }
+router.get('/users/google/login/redirect', passport.authenticate('google'), googleLogin);
+router.get('/users/facebook/login/redirect', passport.authenticate('facebook'), facebookLogin);
 
-      return user.save().then(() => res.json({ user: user.toAuthJSON() }));
-    })
-    .catch(next);
-});
 
-router.post('/users/login', (req, res, next) => {
-  if (!req.body.user.email) {
-    return res.status(422).json({ errors: { email: "can't be blank" } });
-  }
-
-  if (!req.body.user.password) {
-    return res.status(422).json({ errors: { password: "can't be blank" } });
-  }
-  passport.authenticate('local', { session: false }, (
-    err,
-    user,
-    info
-  ) => {
-    if (err) {
-      return next(err);
-    }
-
-    if (user) {
-      return res.json({ user: user.toAuthJSON() });
-    }
-    return res.status(422).json(info);
-  })(req, res, next);
-});
-
-router.post('/users', (req, res, next) => {
-  const user = new User();
-
-  user.username = req.body.user.username;
-  user.email = req.body.user.email;
-  user.setPassword(req.body.user.password);
-
-  user.save()
-    .then(() => res.json({ user: user.toAuthJSON() }))
-    .catch(next);
-});
-
-module.exports = router;
+export default router;
