@@ -4,7 +4,6 @@ export default {
     id: {
       allowNull: false,
       autoIncrement: true,
-      primaryKey: true,
       type: Sequelize.INTEGER
     },
     firstname: {
@@ -18,10 +17,11 @@ export default {
     email: {
       type: Sequelize.STRING,
       allowNull: false,
+      unique: true
     },
     password: {
       type: Sequelize.STRING,
-      allowNull: false,
+      allowNull: true,
     },
     createdAt: {
       allowNull: false,
@@ -32,5 +32,6 @@ export default {
       type: Sequelize.DATE
     }
   }),
+
   down: queryInterface => queryInterface.dropTable('Users')
 };
