@@ -1,11 +1,12 @@
-
 import bcrypt from 'bcryptjs';
+import uuidv4 from 'uuid/v4';
 
 export default {
   up: queryInterface => queryInterface.bulkInsert(
     'Users',
     [
       {
+        userId: uuidv4(),
         firstname: 'Chima',
         lastname: 'Mark',
         email: 'chimamark@example.com',
@@ -14,6 +15,7 @@ export default {
         updatedAt: new Date(),
       },
       {
+        userId: uuidv4(),
         firstname: 'frank',
         lastname: 'chidinma',
         email: 'frankchidinma@example.com',
